requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pandas==2.1.3
requests==2.31.0
python-dotenv==1.0.0
ollama==0.1.7
pydantic==2.5.0
pytest==7.4.3
httpx==0.25.2

# LangChain and RAG dependencies
langchain>=0.1.0
langchain-community>=0.0.10
langchain-core>=0.1.7
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
numpy>=1.24.3
tqdm>=4.66.1

# Demo app
streamlit>=1.28.0